+++ pages/index.tsx
@@ -1,11 +1,11 @@
-import type { NextPage } from 'next'
-import Head from 'next/head'
-import Image from 'next/image'
-import styles from '../styles/Home.module.css'
-import Page from '../src/components/step1/Page'
+import type { NextPage } from "next"
+import Head from "next/head"
+import Image from "next/image"
+import styles from "../styles/Home.module.css"
+import Page from "../src/components/pages/home"
 
 const Home: NextPage = () => {
-  return <Page/>
+  return <Page />
 }
 
 export default Home
